Clear parent DOM node when rendering comments in HTML
because there may be some other content already.

Notably this hides a bug on the Test Run page where we may have a row
with existing comments, which had been expanded previously and then collapsed
(DOM is still there). Then we bulk-update comments, including for the
same row in question, then we expand the row.

Because bulk-update-comments handler *incorrectly* removes the row from
`expandedExecutionIds` next time we expand this row it will refresh
itself and will rended the comment section appended to the already
existing DOM, see #3338.

The correct behavior for bulk-update-comments handler is to append the
new comment to the existing DOM if this row had been expanded before!

Alternate correct behavior is to reload the entire row like we do in
many other places.