Adjust Product & Version display on TestRun page
use the relationship TR->Build as the starting point because TR->Plan
is much more likely to update its Product/Version values in the future.

NOTE: the field `product_version` which is removed from tests doesn't
make sense because neither the underlying forms, not the underlying
models have such a field.