Allow users to reset their email. Fixes #3211
by logging them out and sending a confirmation email, similar to the one
they receive during initial registration because we need to know that
the user is in control of the specified email address.

Administrators which have the `auth.change_user` permission will be able
to reset email addresses for others. Address confirmation still applies!