On New Test Run page always show the Build + button. Refs #3680
NOTES:

If you try assigning a Build from a different product than the
one which is currently selected you will get an error. After an error,
or when creating a Test Run from a Test Plan this page will know exactly
how to prepopulate the Build pop-up in order to avoid error.