Adjust angle bracket icon direction when viewing nested testplans
Fixes #3163

when viewing a test plan which is part of a tree structure all of its
parents appear expended by default and the current TP is highlighted on
the page. This commit adjusts the direction of the existing angle
brackets icons to relfect this state and makes icons for parent test
plans to point down, instead of right!