Update filter logic for Automated=False. Fixes #3841
the boolean expression tc[filterBy] was intended to check whether the
requested field exists, as a way to avoid crashes in the case where we
may have forgotten to update the code. However in the case when trying
to search for manual test cases the raw value of this field is `false`
which caused the entire expression to be evaulated as False and did not
apply the filter.