Speed-up TestCase filtering on the TestPlan view page
by collecting a list of DOM selectors to be queried/shown via jQuery
multiple-selector at the same time instead of calling .show() for each
one of them.

For 1-2 rows this is on-par with the previous implementation, however it is
around 2x faster when the result set contains many rows (say 100 or
1000).

WARNING: when a result set contains hundreds of rows there is still a
noticeable lag on the page. For example if filtering 1000 rows for
is_automated=false, where only 10 rows have the value of True and the
rest 990 have the value of False!

See https://api.jquery.com/multiple-selector/