Add 'Search Test Executions' menu item under SEARCH in navigation bar
This should make it more obvious for users on which page they should get
the information about test executions (because sometimes they get
confused and look for it on the 'Search Test Cases' page)!

This new menu item points to the existing
Telemetry -> Testing -> Execution -> Dashboard
page which displays a formatted table of TestExecution objects and their
statuses.

Refs #3766